+++ Source/Bifrost/Bootstrap/Assemblies/AssemblySpecifiers.cs
@@ -21,7 +21,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Bifrost.Configuration.Assemblies;
-using Bifrost.Extensions;
+using Bifrost.Execution;
 
 namespace Bifrost.Bootstrap.Assemblies
 {
@@ -59,26 +59,52 @@ namespace Bifrost.Bootstrap.Assemblies
                 var specified = false;
                 if (MayReferenceICanSpecifyAssemblies(assembly))
                 {
-                    assembly
-                        .GetTypes()
-                        .Where(t => t.GetInterfaces().Contains(typeof(ICanSpecifyAssemblies)))
-                        .Where(type => type.HasDefaultConstructor())
-                        .ForEach(type =>
-                        {
-                            var specifier = Activator.CreateInstance(type) as ICanSpecifyAssemblies;
-                            specifier.Specify(_assembliesConfiguration);
-                            specified = true;
-                        });
+                    try
+                    {
+                        specified = assembly
+                            .GetTypes()
+                            .Where(t => t.GetInterfaces().Contains(typeof(ICanSpecifyAssemblies)))
+                            .Select(SpecifyFrom)
+                            .ToList()
+                            .Count > 0;
+                    }
+                    catch (ReflectionTypeLoadException e)
+                    {
+                        throw new AssemblySpecificationException(
+                            $"Error while reflecting on the types of {assembly.FullName}. Loader exceptions encountered:\n"+
+                            string.Join("\n", e.LoaderExceptions.Select(l => l.Message).Distinct()));
+                    }
                 }
 
                 return specified;
             }
         }
 
+        ICanSpecifyAssemblies SpecifyFrom(Type type)
+        {
+            try
+            {
+                var specifier = Activator.CreateInstance(type) as ICanSpecifyAssemblies;
+                specifier.Specify(_assembliesConfiguration);
+                return specifier;
+            }
+            catch (MissingMethodException)
+            {
+                throw new AssemblySpecificationException(
+                    $"Could not create instance of type {type.FullName}. It must have a default constructor.");
+            }
+            catch (Exception e)
+            {
+                throw new AssemblySpecificationException(
+                    $"Error while specifying assemblies from {type.FullName}: {e.Message}",
+                    e);
+            }
+        }
+
         static bool MayReferenceICanSpecifyAssemblies(Assembly assembly)
         {
             return assembly.FullName.Contains("Bifrost") ||
-                   assembly.GetReferencedAssemblies().Any(a => a.FullName.Contains("Bifrost"));
+                assembly.GetReferencedAssemblies().Any(a => a.FullName.Contains("Bifrost"));
         }
 #pragma warning restore 1591 // Xml Comments
     }